Issue 6 base UI and navigation creation
* fix git bug commit
* fix previous commit
* delete pycache
* add ignoring to gitignore
