--- portal/navigation/views.py
@@ -1,4 +1,0 @@
-from django.http import HttpResponse
-  
-def index(request):
-    return HttpResponse("Simple page")
