+++ portal/subjects/templates/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>{{ header }}</title>
+</head>
+<body>
+    <h1>{{ text }}</h1><br>
+    <button onclick="location.href='announcements'">Объявления</button> <br>
+    <button onclick="location.href='management'">Администрирование</button><br>
+    <button onclick="location.href='subjects'">Список тем</button><br>
+    <button onclick="location.href='projects'">Проекты</button><br>
+</body>
+</html>
